video-editor-app/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
redis==5.0.1
rq==1.16.2
boto3==1.35.58
requests==2.32.3
pytest==8.2.2
httpx==0.27.2
ruff==0.6.9
black==24.10.0
